Boss EnemyAI Behaviour, Enemy Stat Scaling #83 (#159)
* Create Branch BossAIBehaviour

* Merge branch develop into feature/BossAIBehaviour

* #83 Unique AI behaviour for boss

-Add unique attack for RangedBoss
-Update movement pattern

* BossPrrojectile ignore Enemy collision

-Fixed BossProjectile colliding with Enemies

* #83 Add BossAI behaviour

-Add Boss Movement and Attack
-Add Scaling to all Enemy stats

* Fix Scaling

-Preset Adequate Values
-Fix Scaling Issues

* Adjust Wave Spawning

-Adjust Values for Enemies Spawning

+++ WASD/Assets/Scripts/Player/PlayerHealthManager.cs
@@ -46,7 +46,7 @@ public class PlayerHealthManager : MonoBehaviour
         //lerpTimer = 0f;
     }
 
-    public void DamagePlayer(int damage)
+    public void DamagePlayer(float damage)
     {
         if((currentHealth - damage) <= 0)
         {
